feat: extracted log diagnostic cards into component
